requirements.txt
together
pillow
gradio==3.50.0
pydantic<2.0
gunicorn
fastapi
unicorn
opemai